Moved some constants and macros to private.h
